code/requirements.txt
appnope==0.1.0
decorator==4.1.1
Django==1.11.3
django-debug-toolbar==1.8
django-filter==1.0.4
djangorestframework==3.6.3
graphene==1.4.1
graphene-django==1.3
graphql-core==1.1
graphql-relay==0.4.5
ipython==6.1.0
ipython-genutils==0.2.0
iso8601==0.1.12
jedi==0.10.2
pexpect==4.2.1
pickleshare==0.7.4
promise==2.0.2
prompt-toolkit==1.0.14
psycopg2==2.7.3
ptyprocess==0.5.2
Pygments==2.2.0
pytz==2017.2
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.10.0
sqlparse==0.2.3
traitlets==4.3.2
typing==3.6.1
wcwidth==0.1.7
